- removed conv_stab_fac from parameter list. It is now always 0.5 - splitted the inflow strategie for xfluid and xfluidfluid: ConvStabScaling (xfluid), XFF_ConvStabScaling (xfluidfluid) - XFF_ConvStabScaling has now the four options of inflow stabilizations: one-sided inflow, averaged, and one-sided inflow and averaged combined with max penalty term - new implementation of inflow stabilization terms:   *removed method MHS_Stab_InflowCoercivity, NIT2_Stab_InflowCoercivity and NIT_Stab_InflowCoercivity   *new methods: NIT2_Stab_ConvAveraged, NIT_Stab_ConvAveraged: for calculation of inflow averaged term.   For one-sided inflow stabilization the viscous penalty term is performed in NIT_Stab_ViscCoercivity (with a different penalty parameter)   MSH_Stab_InflowCoercivity: one-sided inflow and averaged inflow terms of mixed/hybrid stress-based methods - bugfix: return value of LINALG::GeneralizedEigen should be double - XFluidlfluid: check if Max reserved Dofs for background dis is not big enough: @Benedikt: you need also something like this.
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19627 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/linalg/linalg_utils.cpp
@@ -684,8 +684,8 @@ void LINALG::SymmetricEigen(Epetra_SerialDenseMatrix& A,
  |  Ax =  lambda Bx via QZ-algorithm (B is singular) and returns the
  |  maximum eigenvalue                              shahmiri  05/13
  *----------------------------------------------------------------------*/
-int LINALG::GeneralizedEigen(Epetra_SerialDenseMatrix& A,
-                             Epetra_SerialDenseMatrix& B)
+double LINALG::GeneralizedEigen(Epetra_SerialDenseMatrix& A,
+                                Epetra_SerialDenseMatrix& B)
 {
   Epetra_SerialDenseMatrix tmpA(A);
   Epetra_SerialDenseMatrix tmpB(B);
